optimize DeleteVersion
